refactor: liskov principle applied and refactoring repositories & services
