+++ src/mappers/UserMap.ts
@@ -1,4 +1,4 @@
-import User from '../models/User';
+import User from '@modules/users/infra/typeorm/entities/User';
 
 export default class UserMap {
     public toDTO({
