+++ src/modules/appointments/dtos/ICreateAppointmentDTO.ts
@@ -1,0 +1,4 @@
+export default interface ICreateAppointmentDTO {
+    provider_id: string;
+    date: Date;
+}
