+++ src/modules/appointments/infra/typeorm/repositories/AppointmentRepository.ts
@@ -1,0 +1,34 @@
+import { getRepository, Repository } from 'typeorm';
+
+import Appointment from '@modules/appointments/infra/typeorm/entities/Appointment';
+import IAppointmentsRepository from '@modules/appointments/repositories/IAppointmentsRepository';
+import ICreateAppointmentDTO from '@modules/appointments/dtos/ICreateAppointmentDTO';
+
+class AppointmentRepository implements IAppointmentsRepository {
+    private ormRepository: Repository<Appointment>;
+
+    constructor() {
+        this.ormRepository = getRepository(Appointment);
+    }
+
+    public async findByDate(date: Date): Promise<Appointment | undefined> {
+        const findAppointment = await this.ormRepository.findOne({
+            where: { date },
+        });
+
+        return findAppointment;
+    }
+
+    public async create({
+        provider_id,
+        date,
+    }: ICreateAppointmentDTO): Promise<Appointment> {
+        const appointment = this.ormRepository.create({ provider_id, date });
+
+        await this.ormRepository.save(appointment);
+
+        return appointment;
+    }
+}
+
+export default AppointmentRepository;
