+++ src/modules/appointments/repositories/IAppointmentsRepository.ts
@@ -1,0 +1,7 @@
+import Appointment from '@modules/appointments/infra/typeorm/entities/Appointment';
+import ICreateAppointmentDTO from '@modules/appointments/dtos/ICreateAppointmentDTO';
+
+export default interface IAppointmentsRepository {
+    create(data: ICreateAppointmentDTO): Promise<Appointment>;
+    findByDate(date: Date): Promise<Appointment | undefined>;
+}
