+++ src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 
-import AuthenticateUserService from '../service/AuthenticateUserService';
-import UserMap from '../mappers/UserMap';
+import AuthenticateUserService from '@modules/users/services/AuthenticateUserService';
+import UserMap from '../../../../../mappers/UserMap';
 
 const sessionsRoutes = Router();
