+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -3,11 +3,11 @@ import { hash } from 'bcryptjs';
 
 import multer from 'multer';
 
-import uploadConfig from '../config/upload';
-import UserMap from '../mappers/UserMap';
-import CreateUserService from '../service/CreateUserService';
-import ensureAuthenticated from '../middlewares/ensureAuthenticated';
-import UpdateUserAvatarService from '../service/UpdateUserAvatarService';
+import uploadConfig from '@config/upload';
+import CreateUserService from '@modules/users/services/CreateUserService';
+import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
+import UpdateUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
+import UserMap from '../../../../../mappers/UserMap';
 
 const usersRoutes = Router();
 const upload = multer(uploadConfig);
