+++ src/modules/users/services/CreateUserService.ts
@@ -1,7 +1,7 @@
 import { getRepository } from 'typeorm';
 
-import User from '../models/User';
-import AppError from '../errors/AppError';
+import User from '@modules/users/infra/typeorm/entities/User';
+import AppError from '@shared/errors/AppError';
 
 interface Request {
     name: string;
