+++ src/shared/infra/http/routes/index.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 
-import appointmentsRoutes from './appointments.routes';
-import usersRoutes from './users.routes';
-import sessionsRoutes from './sessions.routes';
+import usersRoutes from '@modules/users/infra/http/routes/users.routes';
+import appointmentsRoutes from '@modules/appointments/infra/http/routes/appointments.routes';
+import sessionsRoutes from '@modules/users/infra/http/routes/sessions.routes';
 
 const routes = Router();
